auth service added (#1)

+++ implementation/auth/entity.go
@@ -6,6 +6,7 @@ import (
 	"gorm.io/gorm"
 )
 
+// User represents the user model that will be stored in DB and use for internal business logic
 type User struct {
 	Model
 	UserName string `gorm:"user_name"`
@@ -13,8 +14,9 @@ type User struct {
 	Password string `gorm:"password"`
 }
 
+// Model is a helper struct for embedding common fields
 type Model struct {
-	ID        uint           `gorm:"id"`
+	ID        uint64         `gorm:"id"`
 	CreatedAt time.Time      `gorm:"created_at"`
 	UpdatedAt time.Time      `gorm:"updated_at"`
 	DeletedAt gorm.DeletedAt `gorm:"deleted_at"`
